+++ packages/frontend/src/components/Explorer/SavedChartsHeader/index.tsx
@@ -1,10 +1,5 @@
 import { subject } from '@casl/ability';
-import {
-    DashboardTileTypes,
-    FeatureFlags,
-    type ApiError,
-    type PullRequestCreated,
-} from '@lightdash/common';
+import { DashboardTileTypes, FeatureFlags } from '@lightdash/common';
 import {
     ActionIcon,
     Alert,
@@ -22,13 +17,11 @@ import { useDisclosure } from '@mantine/hooks';
 import {
     IconAlertTriangle,
     IconArrowBack,
-    IconArrowRight,
     IconBell,
     IconCheck,
     IconChevronRight,
     IconCirclePlus,
     IconCirclesRelation,
-    IconCodePlus,
     IconCopy,
     IconDatabaseExport,
     IconDots,
@@ -42,10 +35,8 @@ import {
     IconSend,
     IconTrash,
 } from '@tabler/icons-react';
-import { useMutation } from '@tanstack/react-query';
 import { Fragment, useEffect, useMemo, useState, type FC } from 'react';
 import { useBlocker, useLocation, useNavigate, useParams } from 'react-router';
-import { lightdashApi } from '../../../api';
 import { PromotionConfirmDialog } from '../../../features/promotion/components/PromotionConfirmDialog';
 import {
     usePromoteChartDiffMutation,
@@ -60,8 +51,6 @@ import {
 import { SyncModal as GoogleSheetsSyncModal } from '../../../features/sync/components';
 import { useChartViewStats } from '../../../hooks/chart/useChartViewStats';
 import useDashboardStorage from '../../../hooks/dashboard/useDashboardStorage';
-import { useGitIntegration } from '../../../hooks/gitIntegration/useGitIntegration';
-import useToaster from '../../../hooks/toaster/useToaster';
 import { useFeatureFlagEnabled } from '../../../hooks/useFeatureFlagEnabled';
 import { useProject } from '../../../hooks/useProject';
 import {
@@ -101,57 +90,6 @@ enum SpaceType {
 const SpaceTypeLabels = {
     [SpaceType.SharedWithMe]: 'Shared with me',
     [SpaceType.AdminContentView]: 'Public content view',
-};
-
-const createPullRequestForChartFields = async (
-    projectUuid: string,
-    chartUuid: string,
-) =>
-    lightdashApi<any>({
-        url: `/projects/${projectUuid}/git-integration/pull-requests/chart/${chartUuid}/fields`,
-        method: 'GET',
-        body: undefined,
-    });
-
-const useCreatePullRequestForChartFieldsMutation = (
-    projectUuid: string | undefined,
-    chartUuid: string | undefined,
-) => {
-    /* useMutation<GitIntegrationConfiguration, ApiError>(
-        ['git-integration', 'pull-request'],
-        () => createPullRequestForChartFields(projectUuid, chartUuid!),
-
-    );*/
-    const { showToastSuccess, showToastApiError } = useToaster();
-
-    return useMutation<PullRequestCreated, ApiError>(
-        () =>
-            projectUuid && chartUuid
-                ? createPullRequestForChartFields(projectUuid, chartUuid)
-                : Promise.reject(),
-        {
-            mutationKey: ['git-integration', 'pull-request'],
-            retry: false,
-            onSuccess: async (pullRequest) => {
-                showToastSuccess({
-                    title: `Success! Create branch with changes: '${pullRequest.prTitle}'`,
-                    action: {
-                        children: 'Open Pull Request',
-                        icon: IconArrowRight,
-                        onClick: () => {
-                            window.open(pullRequest.prUrl, '_blank');
-                        },
-                    },
-                });
-            },
-            onError: ({ error }) => {
-                showToastApiError({
-                    title: `Failed to create pull request`,
-                    apiError: error,
-                });
-            },
-        },
-    );
 };
 
 type SavedChartsHeaderProps = {
@@ -233,11 +171,6 @@ const SavedChartsHeader: FC<SavedChartsHeaderProps> = ({
         savedChart?.uuid,
     );
     const chartViewStats = useChartViewStats(savedChart?.uuid);
-    const { data: gitIntegration } = useGitIntegration(projectUuid);
-    const createPullRequest = useCreatePullRequestForChartFieldsMutation(
-        projectUuid,
-        savedChart?.uuid,
-    );
     const chartBelongsToDashboard: boolean = !!savedChart?.dashboardUuid;
 
     const hasGoogleDriveEnabled =
@@ -868,18 +801,7 @@ const SavedChartsHeader: FC<SavedChartsHeaderProps> = ({
                                         Google Sheets Sync
                                     </Menu.Item>
                                 ) : null}
-                                {gitIntegration?.enabled && (
-                                    <Menu.Item
-                                        icon={
-                                            <MantineIcon icon={IconCodePlus} />
-                                        }
-                                        onClick={() =>
-                                            createPullRequest.mutate()
-                                        }
-                                    >
-                                        Add custom metrics to dbt project
-                                    </Menu.Item>
-                                )}
+
                                 {userCanManageChart && (
                                     <>
                                         <Menu.Divider />