+++ packages/frontend/src/hooks/echarts/useEchartsCartesianConfig.ts
@@ -1,11 +1,8 @@
 import {
-    CartesianSeriesType,
-    DimensionType,
-    MetricType,
-    TableCalculationType,
-    TimeFrames,
     applyCustomFormat,
     assertUnreachable,
+    CartesianSeriesType,
+    DimensionType,
     formatItemValue,
     formatValueWithExpression,
     friendlyName,
@@ -17,6 +14,7 @@ import {
     getResultValueArray,
     hasFormatting,
     hashFieldReference,
+    hasValidFormatExpression,
     isCompleteLayout,
     isCustomBinDimension,
     isCustomDimension,
@@ -27,7 +25,10 @@ import {
     isPivotReferenceWithValues,
     isTableCalculation,
     isTimeInterval,
+    MetricType,
+    TableCalculationType,
     timeFrameConfigs,
+    TimeFrames,
     type ApiQueryResults,
     type CartesianChart,
     type CustomDimension,
@@ -39,6 +40,7 @@ import {
     type Series,
     type TableCalculation,
 } from '@lightdash/common';
+import { useMantineTheme } from '@mantine/core';
 import dayjs from 'dayjs';
 import {
     type DefaultLabelFormatterCallbackParams,
@@ -584,7 +586,7 @@ type GetPivotSeriesArg = {
 };
 
 const seriesValueFormatter = (item: Item, value: unknown) => {
-    if (isField(item) && item.format) {
+    if (hasValidFormatExpression(item)) {
         return formatValueWithExpression(item.format, value);
     }
 
@@ -1574,6 +1576,8 @@ const useEchartsCartesianConfig = (
         minimal,
     } = useVisualizationContext();
 
+    const theme = useMantineTheme();
+
     const validCartesianConfig = useMemo(() => {
         if (!isCartesianVisualizationConfig(visualizationConfig)) return;
         return visualizationConfig.chartConfig.validConfig;
@@ -1910,6 +1914,9 @@ const useEchartsCartesianConfig = (
                     validCartesianConfig?.eChartsConfig.grid,
                 ),
             },
+            textStyle: {
+                fontFamily: theme?.other.chartFont as string | undefined,
+            },
             // We assign colors per series, so we specify an empty list here.
             color: [],
         }),
@@ -1925,6 +1932,7 @@ const useEchartsCartesianConfig = (
             series,
             sortedResults,
             tooltip,
+            theme?.other?.chartFont,
         ],
     );