+++ packages/frontend/src/sdk/index.tsx
@@ -3,7 +3,8 @@
 
 import { MemoryRouter } from 'react-router';
 
-import EmbedDashboard from '../ee/pages/EmbedDashboard';
+import { FilterOperator } from '@lightdash/common';
+import EmbedDashboard, { type SdkFilter } from '../ee/pages/EmbedDashboard';
 import EmbedProvider from '../ee/providers/Embed/EmbedProvider';
 import ErrorBoundary from '../features/errorBoundary/ErrorBoundary';
 import ChartColorMappingContextProvider from '../hooks/useChartColorConfig/ChartColorMappingContextProvider';
@@ -25,6 +26,7 @@ export {
     EmbedDashboard,
     EmbedProvider,
     ErrorBoundary,
+    FilterOperator,
     FullscreenProvider,
     LIGHTDASH_SDK_INSTANCE_URL_LOCAL_STORAGE_KEY,
     MantineProvider,
@@ -32,4 +34,5 @@ export {
     ReactQueryProvider,
     ThirdPartyServicesProvider,
     TrackingProvider,
+    type SdkFilter,
 };